Refactor conflict detection by table type
